Added user join notification
